restructure dfanalyzer to dftracer.analyzer
