+++ pyproject.toml
@@ -37,14 +37,15 @@ dependencies = [
     "strenum>=0.4",
     "structlog>=25",
     # "dftracer-utils>=0.0.4"
+    "dftracer-utils @ git+https://github.com/rayandrew/dftracer-utils.git@db4125b660367bff7a0c81c6dec93da9be770f51",
 ]
 
 [project.optional-dependencies]
 darshan = ["darshan>=3.4"]
 
 [project.scripts]
-dfanalyzer = "dfanalyzer.__main__:main"
-dfanalyzer-cluster = "dfanalyzer.cluster:main"
+dfanalyzer = "dftracer.analyzer.__main__:main"
+dfanalyzer-cluster = "dftracer.analyzer.cluster:main"
 
 [project.urls]
 documentation = "https://dfanalyzer.readthedocs.io/en/latest/"
@@ -73,4 +74,4 @@ line-length = 119
 quote-style = "preserve"
 
 [tool.spin]
-package = 'dfanalyzer'
+package = 'dftracer.analyzer'
