+++ python/dftracer/analyzer/utils/collection_utils.py
